+++ app/Http/Requests/Api/Admin/Transaction/UpdateRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests\Api\Main\Concert;
+namespace App\Http\Requests\Api\Admin\Transaction;
 
 use App\Http\Requests\ApiRequest;
 
-class StoreRequest extends ApiRequest
+class UpdateRequest extends ApiRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,10 +24,10 @@ class StoreRequest extends ApiRequest
     public function rules()
     {
         return [
-            'company' => ['required','numeric'],
-            'name' => ['required','string','max:255','unique:concerts,name'],
-            'start_at' => ['required','string','date_format:d-m-Y'],
-            'end_at' => ['required','string','date_format:d-m-Y']
+            'concert' => ['nullable','numeric'],
+            'user' => ['nullable','numeric'],
+            'paid_at' => ['nullable','string','date_format:d-m-Y'],
+            'book_at' => ['nullable','string','date_format:d-m-Y']
         ];
     }
 }
